Add WithHTTPClient function for per-request client control
If it's necessary to control the HTTP client on a per-request basis
(instead of using the package-global HTTPClientDefault), there's
currently no safe way to do this except serializing the HTTP requests
and mutating the package global for each.

Add a WithHTTPClient function that sets a key in a context, and have
requestXml read this context value if present when picking the HTTP
client to use.

Signed-off-by: Andrew Dunham <andrew@tailscale.com>